Dante
a novelist whose identity is shaped not only by his production of the Inferno as a vernacular novel, but also by his mythologizations of the woman Beatrice in La Vita Nuova and Paradiso.  Can anyone believe it is a basic accident that the muse was female, and the author male?